backend/requirements.txt
# Backend Dependencies
# Note: If you encounter issues with Python 3.14, use Python 3.11 or 3.12
flask>=3.0.0
flask-cors>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
